after 2-nd code review:  - refactored 'init_db.sql': set BIGINT in column 'price'  - deleted empty line after class declaration 'BookDaoImpl'  - deleted redundant Book book = newBook() in 'findById'

+++ src/main/resources/init_db.sql
@@ -3,6 +3,6 @@ USE book;
 CREATE TABLE `books` (
 `id` BIGINT AUTO_INCREMENT,
 `title` VARCHAR(255),
-`price` BigDecimal,
+`price` BIGINT,
 PRIMARY KEY(`id`)
 );
